+++ kitchen/cookbooks/identity-gitlab/CHANGELOG.md
@@ -1,9 +1,9 @@
-# identity-nessus CHANGELOG
+# identity-gitlab CHANGELOG
 
-This file is used to list changes made in each version of the identity-nessus cookbook.
+This file is used to list changes made in each version of the identity-gitlab cookbook.
 
 ## 0.1.0
-- [your_name] - Initial release of identity-nessus
+- [your_name] - Initial release of identity-gitlab
 
 - - -
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.